Fix warnings from r1948
git-svn-id: svn://rephial.org/angband/trunk@1949 98ddb578-d4ae-11db-a93f-5346e46f259c

+++ src/z-bitflag.c
@@ -49,7 +49,7 @@ bool flag_has_dbg(const bitflag *flags, const size_t size, const int flag, const
 	if (flag_offset >= size)
 	{
 		fprintf(stderr, "flag_has() error %s[%s] Flag[%d] FO[%u] FB[%d] Size[%u]\n",
-		        fi, fl, flag, flag_offset, flag_binary, size);
+		        fi, fl, flag, (unsigned int) flag_offset, flag_binary, (unsigned int) size);
 		fflush(stderr);
 	}
 
@@ -195,7 +195,7 @@ bool flag_on_dbg(bitflag *flags, const size_t size, const int flag, const char *
 	if (flag_offset >= size)
 	{
 		fprintf(stderr, "flag_on() error %s[%s] Flag[%d] FO[%u] FB[%d] Size[%u]\n",
-		        fi, fl, flag, flag_offset, flag_binary, size);
+		        fi, fl, flag, (unsigned int) flag_offset, flag_binary, (unsigned int) size);
 		fflush(stderr);
 	}
